Connect provider dashboard to authorization flow
