+++ package.json
@@ -13,24 +13,14 @@
     "@aws-amplify/auth": "^6.13.0",
     "@aws-amplify/ui-react": "^6.11.1",
     "@headlessui/react": "^2.2.1",
-<<<<<<< HEAD
-<<<<<<< HEAD
     "@radix-ui/react-avatar": "^1.1.10",
     "@radix-ui/react-tabs": "^1.1.12",
-    "aws-amplify": "^6.14.4",
     "bad-words": "^4.0.0",
     "class-variance-authority": "^0.7.1",
     "date-fns": "^4.1.0",
-=======
-=======
     "@hookform/resolvers": "^5.1.1",
-<<<<<<< HEAD
->>>>>>> 3eac1b2 (Add form validation to client registration)
-=======
     "@preline/toggle-password": "^3.1.0",
->>>>>>> 2e3d7ee (Complete functionality to fetch user profile data)
     "aws-amplify": "^6.15.0",
->>>>>>> e55eee9 (Complete reset password functionality)
     "lucide-react": "^0.508.0",
     "next": "15.2.4",
     "openid-client": "^6.5.0",
@@ -39,13 +29,9 @@
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.57.0",
     "react-icons": "^5.5.0",
-<<<<<<< HEAD
     "tailwind": "^4.0.0",
     "tailwind-merge": "^3.3.1",
-    "zod": "^3.24.3"
-=======
     "zod": "^3.25.62"
->>>>>>> 3eac1b2 (Add form validation to client registration)
   },
   "devDependencies": {
     "@aws-amplify/backend": "^1.16.0",
@@ -53,11 +39,8 @@
     "@aws-sdk/client-cognito-identity-provider": "^3.828.0",
     "@eslint/eslintrc": "^3",
     "@tailwindcss/postcss": "^4",
-<<<<<<< HEAD
     "@types/date-fns": "^2.5.3",
-=======
     "@types/aws-lambda": "^8.10.149",
->>>>>>> 94e6bc2 (In progress of adding user to user pool groups)
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
