+++ src/app/page.tsx
@@ -1,13 +1,11 @@
 import "@/app/globals.css";
 
-import { TopNavBar } from "@/components/navbars/top-navbar";
 import HeroSection from "@/components/home/hero-section";
 import AboutSection from "@/components/home/about-section"; // Adjust the path as necessary
 import FeaturesSection from "@/components/home/feature-section";
 import { MeetOurProviders } from "@/components/home/meet-provider-section";
 import { HowItWorks } from "@/components/home/how-it-works";
 import NavBar from "@/components/navbars/navbar";
-import { isAuthenticated } from "@/utils/amplify-server-utils";
 
 export default async function Home() {
   return (
