+++ src/components/dashboards/provider-dashboard.tsx
@@ -3,6 +3,7 @@ import { getCurrentUserServer } from "@/utils/amplify-server-utils";
 import { getUserProfile } from "@/actions/getUserProfile";
 import { redirect } from "next/navigation";
 import { checkIsInGroup } from "@/utils/amplify-server-utils";
+import HomeDashPage from "../provider-dashboard-ui/home-dash-page";
 
 export default async function ProviderDashboard() {
   const isProvider = await checkIsInGroup("Provider");
@@ -15,13 +16,8 @@ export default async function ProviderDashboard() {
 
   return (
     <div>
-      <h1 className="text-3xl font-bold">Provider Dashboard</h1>
-      <p>
-        <strong>Email:</strong> {userProfileData.email}
-      </p>
-      <p>
-        <strong>User Type:</strong> {userProfileData.userType}
-      </p>
+      <h1 className="text-h3-size text-darkest-green my-4">Welcome Jane!</h1>
+      <HomeDashPage />
     </div>
   );
 }
